Updates for new firmware  (#8)
* -Add support for a second button (the BOOT button).  This required modifing the foa functions so that it keys on a void * rather than the IRQ address.
This modification was needed because the IRQ for gpio interrupts are the same so something unique was necessary for mapping to the object class. The
button class now uses the GPIO line itself as the key but the foa function was modified so that whatever the key is, it is simply defined as a
void * so the user can use whatever key desired. (#5)

-adding active state to button class to allow buttons to be active in high or low state

* updating to make LED operations more responsive

* uart2 verified